requirements.txt
fastapi
uvicorn
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
